+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/blocks/EndermanSkullBlock.java
@@ -75,7 +75,7 @@ public class EndermanSkullBlock extends SkullBlock {
     public void animateTick(BlockState state, Level level, BlockPos pos, RandomSource random) {
         if (state.getValue(WATCHED)) {
             ParticleUtil.spawnParticleOnBlockShape(level, pos, ParticleTypes.PORTAL,
-                    UniformInt.of(1, 1+state.getValue(POWER)/2),
+                    UniformInt.of(1, 1 + state.getValue(POWER) / 2),
                     0.5f);
         }
     }