+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/tiles/BookPileBlockTile.java
@@ -1,24 +1,19 @@
 package net.mehvahdjukaar.supplementaries.common.block.tiles;
 
 import com.mojang.blaze3d.vertex.VertexConsumer;
-import com.mojang.serialization.Codec;
 import net.mehvahdjukaar.moonlight.api.block.ItemDisplayTile;
 import net.mehvahdjukaar.moonlight.api.client.model.ExtraModelData;
 import net.mehvahdjukaar.moonlight.api.client.model.IExtraModelDataProvider;
 import net.mehvahdjukaar.moonlight.api.client.model.ModelDataKey;
-import net.mehvahdjukaar.moonlight.api.util.Utils;
-import net.mehvahdjukaar.moonlight.api.util.math.colors.HSLColor;
-import net.mehvahdjukaar.moonlight.api.util.math.colors.RGBColor;
 import net.mehvahdjukaar.supplementaries.client.ModMaterials;
 import net.mehvahdjukaar.supplementaries.client.SpriteCoordinateUnExpander;
-import net.mehvahdjukaar.supplementaries.client.renderers.color.ColorHelper;
 import net.mehvahdjukaar.supplementaries.common.block.ModBlockProperties;
 import net.mehvahdjukaar.supplementaries.common.block.blocks.BookPileBlock;
-import net.mehvahdjukaar.supplementaries.common.items.AntiqueInkItem;
+import net.mehvahdjukaar.supplementaries.common.block.placeable_book.BookType;
+import net.mehvahdjukaar.supplementaries.common.block.placeable_book.PlaceableBookManager;
 import net.mehvahdjukaar.supplementaries.configs.ClientConfigs;
 import net.mehvahdjukaar.supplementaries.configs.CommonConfigs;
 import net.mehvahdjukaar.supplementaries.integration.CompatHandler;
-import net.mehvahdjukaar.supplementaries.integration.CompatObjects;
 import net.mehvahdjukaar.supplementaries.integration.EnchantRedesignCompat;
 import net.mehvahdjukaar.supplementaries.reg.ModRegistry;
 import net.minecraft.client.renderer.MultiBufferSource;
@@ -26,10 +21,11 @@ import net.minecraft.client.renderer.RenderType;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
-import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
-import net.minecraft.util.StringRepresentable;
-import net.minecraft.world.item.*;
+import net.minecraft.world.item.BookItem;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 
@@ -71,7 +67,7 @@ public class BookPileBlockTile extends ItemDisplayTile implements IExtraModelDat
             if (r < 3) it = Items.ENCHANTED_BOOK;
             else if (r < 4) it = Items.WRITABLE_BOOK;
             else it = Items.BOOK;
-            ArrayList<BookColor> col = new ArrayList<>(List.of(BookColor.values()));
+            ArrayList<BookType> col = PlaceableBookManager.getByItem(Items.BOOK.getDefaultInstance());
             books.add(new VisualBook(it.getDefaultInstance(), this.worldPosition, j,
                     col, null));
         }
@@ -129,12 +125,16 @@ public class BookPileBlockTile extends ItemDisplayTile implements IExtraModelDat
     @Override
     public void updateClientVisualsOnLoad() {
         this.books.clear();
-        List<BookColor> colors = new ArrayList<>(ClientConfigs.Tweaks.BOOK_COLORS.get());
-        for (int i = 0; i < 4; i++) {
-            ItemStack stack = this.getItem(i);
+        List<BookType> colors = new ArrayList<>();
+        for (var v : ClientConfigs.Tweaks.BOOK_COLORS.get()) {
+            BookType byName = PlaceableBookManager.getByName(v);
+            if (!colors.contains(byName)) colors.add(byName);
+        }
+        for (int index = 0; index < 4; index++) {
+            ItemStack stack = this.getItem(index);
             if (stack.isEmpty()) break;
-            BookColor last = i == 0 ? null : this.books.get(i - 1).color;
-            this.books.add(i, new VisualBook(stack, this.worldPosition, i, colors, last));
+            BookType last = index == 0 ? null : this.books.get(index - 1).type;
+            this.books.add(index, new VisualBook(stack, this.worldPosition, index, colors, last));
         }
 
         if (books.isEmpty()) {
@@ -152,52 +152,35 @@ public class BookPileBlockTile extends ItemDisplayTile implements IExtraModelDat
     }
 
     //only client
-
-    //TODO: refactor pls
     public static class VisualBook {
-        private final float angle;
-        private final @Nullable BookColor color;
+        private final float yAngle;
+        private final BookType type;
         private final ItemStack stack;
-        private final boolean isEnchanted;
 
-        public VisualBook(ItemStack stack, BlockPos pos, int index, List<BookColor> colors, @Nullable BookColor lastColor) {
-            this.stack = stack;
+        public VisualBook(ItemStack bookStack, BlockPos pos, int index, List<BookType> colors, @Nullable BookType lastColor) {
+            this.stack = bookStack;
             Random rand = new Random(pos.asLong());
             for (int j = 0; j < index; j++) rand.nextInt();
-            Item item = stack.getItem();
-            this.angle = (float) (rand.nextInt(32) * Math.PI / 16);
+            Item item = bookStack.getItem();
+            this.yAngle = (float) (rand.nextInt(32) * Math.PI / 16);
 
-            if( item == CompatObjects.GENE_BOOK){
-                this.isEnchanted = false;
-                this.color = BookColor.GENE;
-            }
-            else if (item instanceof BookItem) {
+            if (item instanceof BookItem) {
                 if (lastColor == null) {
-                    this.color = colors.get(rand.nextInt(colors.size()));
+                    this.type = colors.get(rand.nextInt(colors.size()));
                 } else {
-                    List<BookColor> c = colors.stream().filter(b -> b.looksGoodNextTo(lastColor)).toList();
-                    this.color = c.get(rand.nextInt(c.size()));
+                    List<BookType> c = colors.stream().filter(b -> b.looksGoodNextTo(lastColor)).toList();
+                    this.type = c.get(rand.nextInt(c.size()));
                 }
-                colors.remove(this.color);
-                this.isEnchanted = false;
-            } else if (Utils.getID(item).getNamespace().equals("inspirations")) {
-                String colName = Utils.getID(item).getPath().replace("_book", "");
-                this.color = BookColor.byName(colName);
-                this.isEnchanted = false;
-            } else if (BookPileBlock.isWrittenBook(item)) {
-                if (item instanceof WrittenBookItem) {
-                    this.color = AntiqueInkItem.hasAntiqueInk(stack) ? BookColor.TATTERED : BookColor.WRITTEN;
-                } else this.color = BookColor.AND_QUILL;
-                this.isEnchanted = false;
+                colors.remove(this.type);
             } else {
-                this.color = BookPileBlock.isQuarkTome(item) ? BookColor.TOME : BookColor.ENCHANTED;
-                this.isEnchanted = true;
+                var possibleTypes = PlaceableBookManager.getByItem(bookStack);
+                this.type = possibleTypes.get(rand.nextInt(possibleTypes.size()));
             }
         }
 
         @SuppressWarnings("ConstantConditions")
         public VertexConsumer getBuilder(MultiBufferSource buffer) {
-            if (this.isEnchanted && ClientConfigs.Tweaks.BOOK_GLINT.get()) {
+            if (this.type.hasGlint() && ClientConfigs.Tweaks.BOOK_GLINT.get()) {
                 VertexConsumer foilBuilder = null;
                 if (CompatHandler.ENCHANTEDBOOKREDESIGN) {
                     foilBuilder = EnchantRedesignCompat.getBookColoredFoil(this.stack, buffer);
@@ -212,137 +195,44 @@ public class BookPileBlockTile extends ItemDisplayTile implements IExtraModelDat
         }
 
         public float getAngle() {
-            return angle;
+            return yAngle;
         }
 
-        public boolean isEnchanted() {
-            return isEnchanted;
-        }
-
-        public BookColor getColor() {
-            return color;
+        public BookType getType() {
+            return type;
         }
     }
 
-    public static final List<BookColor> DEFAULT_COLORS = List.of(BookColor.BROWN, BookColor.ORANGE, BookColor.YELLOW, BookColor.RED,
-            BookColor.DARK_GREEN, BookColor.LIME, BookColor.TEAL, BookColor.BLUE, BookColor.PURPLE);
+    public static final List<String> DEFAULT_COLORS = List.of("brown", "orange", "yellow",
+            "red", "dark_green", "lime", "teal", "blue", "purple");
 
-    public enum BookColor implements StringRepresentable {
-        BROWN(DyeColor.BROWN, 1),
-        WHITE(DyeColor.WHITE, 1),
-        BLACK(DyeColor.BLACK, 1),
-        LIGHT_GRAY(DyeColor.LIGHT_GRAY),
-        GRAY(DyeColor.GRAY),
-        ORANGE(DyeColor.ORANGE),
-        YELLOW(DyeColor.YELLOW),
-        LIME(DyeColor.LIME),
-        DARK_GREEN("green", 0x2fc137),
-        TEAL("cyan", 0x16ecbf),
-        LIGHT_BLUE(DyeColor.LIGHT_BLUE),
-        BLUE(DyeColor.BLUE),
-        PURPLE(DyeColor.PURPLE),
-        MAGENTA(DyeColor.MAGENTA),
-        PINK(DyeColor.PINK),
-        RED(DyeColor.RED),
-        ENCHANTED("enchanted", 0, 1),
-        AND_QUILL("and_quill", 0, 1),
-        WRITTEN("written", 0, 1),
-        TOME("tome", 0, 1),
-        TATTERED("tattered", 0, 1),
-        GENE("gene", 0, 1);
-
-        private final String name;
-        private final float hue;
-        private final float angle;
-
-        BookColor(String s, int rgb, float angle) {
-            this.name = s;
-            var col = new RGBColor(rgb).asHSL();
-            this.hue = col.hue();
-            if (angle < 0) this.angle = getAllowedHueShift(col);
-            else this.angle = Math.max(1, angle);
+    public record BooksList(List<VisualBook> books) {
+        public BooksList() {
+            this(new ArrayList<>());
         }
-
-        BookColor(DyeColor color, float angle) {
-            this(color.getName(), ColorHelper.pack(color.getTextureDiffuseColors()), angle);
-        }
-
-        BookColor(String name, int color) {
-            this(name, color, -1);
-        }
-
-        BookColor(DyeColor color) {
-            this(color.getName(), ColorHelper.pack(color.getTextureDiffuseColors()), -1);
-        }
-
-        public static BookColor byName(String name) {
-            for (BookColor c : values()) {
-                if (c.name.equals(name)) {
-                    return c;
-                }
-            }
-            return BROWN;
-        }
-
-        public boolean looksGoodNextTo(BookColor other) {
-            float diff = Math.abs(Mth.degreesDifference(this.hue * 360, other.hue * 360) / 360);
-            return diff < (other.angle + this.angle) / 2f;
-        }
-
-        //could even just use distance
-        private float getAllowedHueShift(HSLColor color) {
-            float l = color.lightness();
-            float s = ColorHelper.normalizeSaturation(color.saturation(), l);
-            float minAngle = 90 / 360f;
-            float addAngle = 65 / 360f;
-            float distLightSq = 2;//(s * s) + (1 - l) * (1 - l);
-            float distDarkSq = ((s * s) + (l * l));
-            float distSq = Math.min(1, Math.min(distDarkSq, distLightSq));
-            return minAngle + (1 - distSq) * addAngle;
-        }
-
-        public static BookColor rand(Random r) {
-            return values()[r.nextInt(values().length)];
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public String getSerializedName() {
-            return getName();
-        }
-
-        public static final Codec<BookColor> CODEC = StringRepresentable.fromEnum(BookColor::values);
-    }
-
-    public static class BooksList {
-
-        private final List<VisualBook> list = new ArrayList<>();
 
         public void add(VisualBook visualBook) {
-            list.add(visualBook);
+            books.add(visualBook);
         }
 
         public void add(int i, VisualBook visualBook) {
-            list.add(i, visualBook);
-        }
-
-        public boolean isEmpty() {
-            return list.isEmpty();
+            books.add(i, visualBook);
         }
 
         public void clear() {
-            list.clear();
+            books.clear();
+        }
+
+        public boolean isEmpty() {
+            return books.isEmpty();
         }
 
         public VisualBook get(int i) {
-            return list.get(i);
+            return books.get(i);
         }
 
         public int size() {
-            return list.size();
+            return books.size();
         }
     }
 }